+++ include/core/parameters.h
@@ -695,7 +695,11 @@ namespace Parameters
   {
     bool use_default_stabilization;
 
-    bool heat_transfer_stabilization;
+    bool heat_transfer_dcdd_stabilization;
+
+    /// Boolean indicating if the DCDD stabilization term for the VOF phase
+    /// fraction should be assembled (@p true) or not (@p false).
+    bool vof_dcdd_stabilization;
 
     // pressure scaling factor used to facilitate the linear solving when
     // pressure and velocity have very different scales