+++ source/core/parameters.cc
@@ -709,10 +709,18 @@ namespace Parameters
         "Choices are <pspg_supg|gls|grad_div>.");
 
       prm.declare_entry(
-        "heat transfer stabilization",
+        "heat transfer dcdd stabilization",
         "false",
         Patterns::Bool(),
-        "Apply discontinuity-capturing directional dissipation stabilization term on heat transfer <true|false>");
+        "Apply Discontinuity-Capturing Directional Dissipation (DCDD) "
+        "stabilization term on heat transfer <true|false>");
+
+      prm.declare_entry(
+        "vof dcdd stabilization",
+        "true",
+        Patterns::Bool(),
+        "Apply Discontinuity-Capturing Directional Dissipation (DCDD) "
+        "stabilization term on the VOF phase fraction <true|false>");
 
       prm.declare_entry(
         "pressure scaling factor",
@@ -742,7 +750,10 @@ namespace Parameters
       else
         throw(std::runtime_error("Invalid stabilization strategy"));
 
-      heat_transfer_stabilization = prm.get_bool("heat transfer stabilization");
+      // DCDD stabilization activation parameters
+      heat_transfer_dcdd_stabilization =
+        prm.get_bool("heat transfer dcdd stabilization");
+      vof_dcdd_stabilization = prm.get_bool("vof dcdd stabilization");
 
       pressure_scaling_factor = prm.get_double("pressure scaling factor");
     }